Refactored away HashArray
Removed a bunch of unused code and made SnakList simpler.

SnakList can be improved further in follow ups:

* Constructor can hide extending of ArrayObject
* Perhaps can use composition of ArrayObject instead of inheritance
* removeSnakHash -> removeSnakByHash
* ...
