--- src/HashArray.php
@@ -1,329 +1,0 @@
-<?php
-
-namespace Wikibase\DataModel;
-
-use ArrayObject;
-use Hashable;
-use InvalidArgumentException;
-use Traversable;
-
-/**
- * Generic array object with lookups based on hashes of the elements.
- *
- * Elements need to implement Hashable.
- *
- * Note that by default the getHash method uses @see MapValueHashesr
- * which returns a hash based on the contents of the list, regardless
- * of order and keys.
- *
- * Also note that if the Hashable elements are mutable, any modifications
- * made to them via their mutator methods will not cause an update of
- * their associated hash in this array.
- *
- * @since 0.1
- *
- * @license GPL-2.0+
- * @author Jeroen De Dauw < jeroendedauw@gmail.com >
- */
-abstract class HashArray extends ArrayObject {
-
-	/**
-	 * Maps element hashes to their offsets.
-	 *
-	 * @since 0.1
-	 *
-	 * @var array [ element hash (string) => element offset (string|int) ]
-	 */
-	protected $offsetHashes = [];
-
-	/**
-	 * @var integer
-	 */
-	protected $indexOffset = 0;
-
-	/**
-	 * Returns the name of an interface/class that the element should implement/extend.
-	 *
-	 * @since 0.4
-	 *
-	 * @return string
-	 */
-	abstract public function getObjectType();
-
-	/**
-	 * @see ArrayObject::__construct
-	 *
-	 * @param array|Traversable|null $input
-	 * @param int $flags
-	 * @param string $iteratorClass
-	 *
-	 * @throws InvalidArgumentException
-	 */
-	public function __construct( $input = null, $flags = 0, $iteratorClass = 'ArrayIterator' ) {
-		parent::__construct( [], $flags, $iteratorClass );
-
-		if ( $input !== null ) {
-			if ( !is_array( $input ) && !( $input instanceof Traversable ) ) {
-				throw new InvalidArgumentException( '$input must be an array or Traversable' );
-			}
-
-			foreach ( $input as $offset => $value ) {
-				$this->offsetSet( $offset, $value );
-			}
-		}
-	}
-
-	/**
-	 * Finds a new offset for when appending an element.
-	 * The base class does this, so it would be better to integrate,
-	 * but there does not appear to be any way to do this...
-	 *
-	 * @return integer
-	 */
-	protected function getNewOffset() {
-		while ( $this->offsetExists( $this->indexOffset ) ) {
-			$this->indexOffset++;
-		}
-
-		return $this->indexOffset;
-	}
-
-	/**
-	 * Gets called before a new element is added to the ArrayObject.
-	 *
-	 * At this point the index is always set (ie not null) and the
-	 * value is always of the type returned by @see getObjectType.
-	 *
-	 * Should return a boolean. When false is returned the element
-	 * does not get added to the ArrayObject.
-	 *
-	 * @since 0.1
-	 *
-	 * @param int|string $index
-	 * @param Hashable $hashable
-	 *
-	 * @return bool
-	 */
-	protected function preSetElement( $index, $hashable ) {
-		$hash = $hashable->getHash();
-
-		$hasHash = $this->hasElementHash( $hash );
-
-		if ( $hasHash ) {
-			return false;
-		} else {
-			$this->offsetHashes[$hash] = $index;
-
-			return true;
-		}
-	}
-
-	/**
-	 * Returns if there is an element with the provided hash.
-	 *
-	 * @since 0.1
-	 *
-	 * @param string $elementHash
-	 *
-	 * @return bool
-	 */
-	public function hasElementHash( $elementHash ) {
-		return array_key_exists( $elementHash, $this->offsetHashes );
-	}
-
-	/**
-	 * Returns if there is an element with the same hash as the provided element in the list.
-	 *
-	 * @since 0.1
-	 *
-	 * @param Hashable $element
-	 *
-	 * @return bool
-	 */
-	public function hasElement( Hashable $element ) {
-		return $this->hasElementHash( $element->getHash() );
-	}
-
-	/**
-	 * Removes the element with the hash of the provided element, if there is such an element in the list.
-	 *
-	 * @since 0.1
-	 *
-	 * @param Hashable $element
-	 */
-	public function removeElement( Hashable $element ) {
-		$this->removeByElementHash( $element->getHash() );
-	}
-
-	/**
-	 * Removes the element with the provided hash, if there is such an element in the list.
-	 *
-	 * @since 0.1
-	 *
-	 * @param string $elementHash
-	 */
-	public function removeByElementHash( $elementHash ) {
-		if ( $this->hasElementHash( $elementHash ) ) {
-			$offset = $this->offsetHashes[$elementHash];
-			$this->offsetUnset( $offset );
-		}
-	}
-
-	/**
-	 * Adds the provided element to the list if there is no element with the same hash yet.
-	 *
-	 * @since 0.1
-	 *
-	 * @param Hashable $element
-	 *
-	 * @return bool Indicates if the element was added or not.
-	 */
-	public function addElement( Hashable $element ) {
-		$append = !$this->hasElementHash( $element->getHash() );
-
-		if ( $append ) {
-			$this->append( $element );
-		}
-
-		return $append;
-	}
-
-	/**
-	 * Returns the element with the provided hash or false if there is no such element.
-	 *
-	 * @since 0.1
-	 *
-	 * @param string $elementHash
-	 *
-	 * @return mixed|bool
-	 */
-	public function getByElementHash( $elementHash ) {
-		if ( $this->hasElementHash( $elementHash ) ) {
-			$offset = $this->offsetHashes[$elementHash];
-			return $this->offsetGet( $offset );
-		}
-
-		return false;
-	}
-
-	/**
-	 * @see ArrayObject::offsetUnset
-	 *
-	 * @since 0.1
-	 *
-	 * @param mixed $index
-	 */
-	public function offsetUnset( $index ) {
-		if ( $this->offsetExists( $index ) ) {
-			/**
-			 * @var Hashable $element
-			 */
-			$element = $this->offsetGet( $index );
-
-			$hash = $element->getHash();
-
-			unset( $this->offsetHashes[$hash] );
-
-			parent::offsetUnset( $index );
-		}
-	}
-
-	/**
-	 * @see ArrayObject::append
-	 *
-	 * @param mixed $value
-	 */
-	public function append( $value ) {
-		$this->setElement( null, $value );
-	}
-
-	/**
-	 * @see ArrayObject::offsetSet()
-	 *
-	 * @param mixed $index
-	 * @param mixed $value
-	 */
-	public function offsetSet( $index, $value ) {
-		$this->setElement( $index, $value );
-	}
-
-	/**
-	 * Returns if the provided value has the same type as the elements
-	 * that can be added to this ArrayObject.
-	 *
-	 * @param mixed $value
-	 *
-	 * @return bool
-	 */
-	protected function hasValidType( $value ) {
-		$class = $this->getObjectType();
-		return $value instanceof $class;
-	}
-
-	/**
-	 * Method that actually sets the element and holds
-	 * all common code needed for set operations, including
-	 * type checking and offset resolving.
-	 *
-	 * If you want to do additional indexing or have code that
-	 * otherwise needs to be executed whenever an element is added,
-	 * you can overload @see preSetElement.
-	 *
-	 * @param mixed $index
-	 * @param mixed $value
-	 *
-	 * @throws InvalidArgumentException
-	 */
-	protected function setElement( $index, $value ) {
-		if ( !$this->hasValidType( $value ) ) {
-			$type = is_object( $value ) ? get_class( $value ) : gettype( $value );
-
-			throw new InvalidArgumentException( '$value must be an instance of ' . $this->getObjectType() . '; got ' . $type );
-		}
-
-		if ( $index === null ) {
-			$index = $this->getNewOffset();
-		}
-
-		if ( $this->preSetElement( $index, $value ) ) {
-			parent::offsetSet( $index, $value );
-		}
-	}
-
-	/**
-	 * @see Serializable::serialize
-	 *
-	 * @return string
-	 */
-	public function serialize() {
-		return serialize( [
-			'data' => $this->getArrayCopy(),
-			'index' => $this->indexOffset,
-		] );
-	}
-
-	/**
-	 * @see Serializable::unserialize
-	 *
-	 * @param string $serialized
-	 */
-	public function unserialize( $serialized ) {
-		$serializationData = unserialize( $serialized );
-
-		foreach ( $serializationData['data'] as $offset => $value ) {
-			// Just set the element, bypassing checks and offset resolving,
-			// as these elements have already gone through this.
-			parent::offsetSet( $offset, $value );
-		}
-
-		$this->indexOffset = $serializationData['index'];
-	}
-
-	/**
-	 * @return bool
-	 */
-	public function isEmpty() {
-		return !$this->getIterator()->valid();
-	}
-
-}
