--- tests/fixtures/HashArrayWithoutDuplicates.php
@@ -1,18 +1,0 @@
-<?php
-
-namespace Wikibase\DataModel\Fixtures;
-
-use Hashable;
-use Wikibase\DataModel\HashArray;
-
-/**
- * @license GPL-2.0+
- * @author Jeroen De Dauw < jeroendedauw@gmail.com >
- */
-class HashArrayWithoutDuplicates extends HashArray {
-
-	public function getObjectType() {
-		return Hashable::class;
-	}
-
-}
